backend/requirements.txt
Flask
pyserial
flask-cors 